Pull out BookAction
