+++ GrailDiary/DocumentList/BookCollectionViewSnapshotBuilder.swift
@@ -1,10 +1,4 @@
-//
-//  BookCollectionViewSnapshot.swift
-//  BookCollectionViewSnapshot
-//
-//  Created by Brian Dewey on 8/29/21.
-//  Copyright © 2021 Brian's Brain. All rights reserved.
-//
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import Foundation
 import UIKit
@@ -82,4 +76,3 @@ struct BookCollectionViewSnapshotBuilder: Equatable {
     return bookSection
   }
 }
-
