+++ GrailDiary/DocumentList/BookSection.swift
@@ -1,10 +1,4 @@
-//
-//  BookSection.swift
-//  BookSection
-//
-//  Created by Brian Dewey on 8/29/21.
-//  Copyright © 2021 Brian's Brain. All rights reserved.
-//
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import Foundation
 
@@ -22,4 +16,3 @@ enum BookSection {
   /// The sections that hold books.
   static let bookSections: [BookSection] = [.currentlyReading, .wantToRead, .read]
 }
-
