+++ GrailDiary/DocumentList/BookViewProperties.swift
@@ -1,10 +1,4 @@
-//
-//  BookViewProperties.swift
-//  BookViewProperties
-//
-//  Created by Brian Dewey on 8/29/21.
-//  Copyright © 2021 Brian's Brain. All rights reserved.
-//
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import Foundation
 
@@ -96,4 +90,3 @@ struct BookViewProperties: Hashable {
       (rhsRating, rhs.noteProperties.creationTimestamp, rhs.author, rhs.noteProperties.title, rhs.noteProperties.modifiedTimestamp)
   }
 }
-
