+++ GrailDiary/DocumentList/ClearBackgroundCell.swift
@@ -1,10 +1,4 @@
-//
-//  ClearBackgroundCell.swift
-//  ClearBackgroundCell
-//
-//  Created by Brian Dewey on 8/29/21.
-//  Copyright © 2021 Brian's Brain. All rights reserved.
-//
+// Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import UIKit
 
@@ -21,5 +15,3 @@ internal final class ClearBackgroundCell: UICollectionViewListCell {
     self.backgroundConfiguration = backgroundConfiguration
   }
 }
-
-
